Replace most of the release process with a workflow
This does all the things you need to do to make a release - except for
the tagging - in one job. These things all happens in one single PR,
and they happen before you need to tag the release so you can decide
to tag it or not.

I'm happy to report that I only resorted to sed in one single place.