Change from 'password' to 'passwordHash' in chart
This is a breaking change with previous versions of the chart (new
version 2.0.0)

Given that weave-gitops is almost always going to be used as part of a
flux-enabled system the secret that contains the adminUser should,
ideally, be managed in-code. Rather than encourage users to store
passwords in plain-text by applying hashing for them they should store
the bcrypt-hashed password (which is much less vulnerable).

cf. https://github.com/OWASP/CheatSheetSeries/blob/master/cheatsheets/Password_Storage_Cheat_Sheet.md