ui: Make CoreClientContext monkeypatch work
This seems to just never have worked? Whenever CoreClientContext was
initialised, it just ignored the API and returned an empty
object. This then means that anything trying to use the API will fail.

The fix here is to use `getOwnPropertyNames` instead - static methods
aren't iterable, so we have to use something more dramatic. However,
that also returns builtins like `prototype`, `name`, `length`, and we
don't want to monkeypatch those, so just monkeypatch methods.

I've written a test that fails with the old code and passes with the
new code.